+++ src/main/java/com/enyoi/apimesaayuda/aplicacion/services/impl/SolicitudesService.java
@@ -59,8 +59,7 @@ public class SolicitudesService implements ISolicitudesService {
         Solicitudes solicitudes;
         if (solicitudesOptional.isPresent()) {
             solicitudes = solicitudesOptional.get();
-            SolicitudesDTO solicitudesDTO = modelMapper.map(solicitudes, SolicitudesDTO.class);
-            return solicitudesDTO;
+            return modelMapper.map(solicitudes, SolicitudesDTO.class);
         } else {
             throw new NotDataFound(NOEXISTENDATOS);
         }
@@ -123,7 +122,14 @@ public class SolicitudesService implements ISolicitudesService {
 
     @Override
     public Page<SolicitudesDTO> findByFechaFinalizadoBetween(Date fechaInicio, Date fechaFin, int page, int size, String columnFilter, Sort.Direction direction) {
-        return null;
+        Pageable pageable = PageRequest.of(page, size, Sort.by(direction, columnFilter));
+        List<SolicitudesDTO> list = solicitudesRepository.findByFechaFinalizadoBetween(fechaInicio, fechaFin, pageable)
+                .stream()
+                .map(solicitudes -> modelMapper.map(solicitudes, SolicitudesDTO.class))
+                .collect(Collectors.toList());
+        final int start = (int) pageable.getOffset();
+        final int end = Math.min((start + pageable.getPageSize()), list.size());
+        return new PageImpl<>(list.subList(start, end), pageable, list.size());
     }
 
     @Override
@@ -180,16 +186,18 @@ public class SolicitudesService implements ISolicitudesService {
             solicitudesGuardar.setTitulo(actualizarSolicitudesRequest.getTitulo());
             Dependencias dependencias = dependenciasRepository.findById(actualizarSolicitudesRequest.getDependenciasId())
                     .orElseThrow(() -> new NotDataFound(NOEXISTENDATOS));
+            solicitudesGuardar.setDependenciasId(dependencias);
             EstadosSolicitud estadosSolicitud = estadosSolicitudRepository.findById(actualizarSolicitudesRequest.getEstadoId())
                     .orElseThrow(() -> new NotDataFound(NOEXISTENDATOS));
+            solicitudesGuardar.setEstadoId(estadosSolicitud);
             Usuarios usuarios = usuariosRepository.findById(actualizarSolicitudesRequest.getSolicitanteId())
                     .orElseThrow(() -> new NotDataFound(NOEXISTENDATOS));
+            solicitudesGuardar.setSolicitanteId(usuarios);
             TiposSolicitud tiposSolicitud = tiposSolicitudRepository.findById(actualizarSolicitudesRequest.getTipoSolicitudId())
                     .orElseThrow(() -> new NotDataFound(NOEXISTENDATOS));
-
+            solicitudesGuardar.setTipoSolicitudId(tiposSolicitud);
             solicitudesGuardar = solicitudesRepository.save(solicitudesGuardar);
             return modelMapper.map(solicitudesGuardar, SolicitudesDTO.class);
-
         } else {
             throw new NotDataFound("codigo de solicitud no existe");
         }
@@ -199,10 +207,13 @@ public class SolicitudesService implements ISolicitudesService {
     public String eliminar(Long id) {
         Optional<Solicitudes> solicitudesOptional = Optional.ofNullable(solicitudesRepository.findById(id)
                 .orElseThrow(() -> new NotDataFound(" Solicitud No existe: " + id)));
+        if (solicitudesOptional.isPresent()) {
+            solicitudesRepository.deleteById(id);
 
-        solicitudesRepository.deleteById(id);
-
-        return solicitudesOptional.get() + "Eliminado con Exito";
+            return modelMapper.map(solicitudesOptional.get(), SolicitudesDTO.class).getId() + "Eliminado con Exito";
+        } else {
+            throw new NotDataFound(NOEXISTENDATOS);
+        }
     }
 
 }
